refactor categories folder
