+++ frontend/src/components/CategoriesNav.jsx
