+++ frontend/src/pages/Home.jsx
@@ -1,15 +1,12 @@
-import GenericBtns from "../components/UI/GenericBtns";
-import Featured from "../components/featured/Featured";
-import CategoriesNav from "../components/categories/CategoriesNav";
-import { userStore } from "../features/userStore";
+import CategoriesNav from "../components/CategoriesNav";
+import FeaturedListWrapper from "../components/featured/FeaturedListWrapper";
 
 const Home = () => {
-  const user = userStore((state) => state.user);
   return (
     <>
       <section className="flex flex-col items-start justify-start col-start-3 col-span-full">
         <CategoriesNav />
-        <Featured />
+        <FeaturedListWrapper />
       </section>
     </>
   );
